registration ok

+++ TaxiBackOffice/src/main/java/shashi/uomtrust/ac/mu/BackOfficeUser.java
@@ -13,17 +13,17 @@ public class BackOfficeUser extends User {
 		
 	private static final long serialVersionUID = -1460047265506653280L;
 
-	private Long id;
+	private Integer id;
 	
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 	
-	 public BackOfficeUser(Long id, String email, String password, boolean enabled, Collection<? extends GrantedAuthority> authorities ) {
+	 public BackOfficeUser(Integer id, String email, String password, boolean enabled, Collection<? extends GrantedAuthority> authorities ) {
 	    	super(email, password, enabled, true, true, true, authorities);
 			this.id = id;
 	 }
