+++ TaxiBackOffice/src/main/java/shashi/uomtrust/ac/mu/service/AccountService.java
@@ -4,7 +4,7 @@ import shashi.uomtrust.ac.mu.entity.Account;
 
 public interface AccountService {
 
-	public Account findByAccountId(Long accountId);
+	public Account findByAccountId(Integer accountId);
 	public Account findByEmail(String email);
 	public Account saveAccount(Account account);
 	public boolean checkAdminLogin(Account account);
